+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Fix OpenAL deadlocks
 - Change frame sleep logic to get fps closer to target fps
+- Auto-repeat bug, which forced to disable xlib auto-repeat during the game execution
 
 ## [1.3.0] - 2018-07-30
 ### Added